Fixes GH-3: Random string generation methods should be consistent.
This is an overhaul of the entire string randomization infrastructure. The result is StringGenerator which can provide strings
of arbitrary codepoint or codeunit length (random or fixed length if min/max range is the same). Different generator strategies
have been implemented or refactored (ascii, codepoint set, unicode, realistic unicode).

This commit is not backwards compatible, update your code.